VarsityPro-Assignment-Task/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
distlib==0.3.6
Django==3.0.5
django-admin-searchable-dropdown==1.2
django-countries==7.5.1
django-extensions==3.2.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
feedparser==6.0.10
filelock==3.8.0
graphviz==0.20.1
html5lib==1.1
idna==3.4
lxml==4.9.2
myenv==201902.7
mysqlclient==2.1.1
oscrypto==1.3.0
Pillow==9.4.0
pipenv==2023.3.20
platformdirs==2.5.2
pycparser==2.21
pydot==1.4.2
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2022.5
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.21.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xhtml2pdf==0.2.8
